refactor: move database/dto/* to corresponding resources
