+++ src/albums/dto/update-album.dto.ts
