+++ src/artists/dto/create-artist.dto.ts
