+++ src/tracks/tracks.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { DatabaseService } from 'src/database/database.service';
-import { CreateTrackDto } from 'src/database/services/dto/tracks/create-track.dto';
-import { UpdateTrackDto } from 'src/database/services/dto/tracks/update-track.dto';
+import { CreateTrackDto } from 'src/tracks/dto/create-track.dto';
+import { UpdateTrackDto } from 'src/tracks/dto/update-track.dto';
 import { ID } from 'src/database/types/models';
 
 @Injectable()
