+++ src/users/users.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { DatabaseService } from 'src/database/database.service';
-import { CreateUserDto } from 'src/database/services/dto/users/create-user.dto';
-import { UpdateUserPasswordDto } from 'src/database/services/dto/users/update-user-password.dto';
+import { CreateUserDto } from 'src/users/dto/create-user.dto';
+import { UpdateUserPasswordDto } from 'src/users/dto/update-user-password.dto';
 import { ID } from 'src/database/types/models';
 
 @Injectable()
